feat: Show PR Data scroll bar only if it is not empty
Currently, when you type in a correct repo, and hit search, the loading icon will have a scroll
bar next to it. This also happens again if, after a successful repo name search, you type
in a wrong repo name, and try to type something again into the search bar. You will notice that
a scroll bar on an empty DIV will appear.

This commit fixes this issue by only adding the className="pullRequestContainer" to the div
only if the pr data array is not empty.